Setup eslint and prettier for the workspace
This project is going to use eslint + prettier to
control the look of the code and to minimize merge
conflicts in the future. The airbnb code convention
is the main source for our code.
