+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+  "singleQuote": true,
+  "useTabs": true,
+  "tabWidth": 1
+}
